+++ pandatool/src/pstatserver/pStatThreadData.h
@@ -49,19 +49,19 @@ public:
   bool has_frame(int frame_number) const;
   const PStatFrameData &get_frame(int frame_number) const;
 
-  float get_latest_time() const;
-  float get_oldest_time() const;
-  const PStatFrameData &get_frame_at_time(float time) const;
-  int get_frame_number_at_time(float time, int hint = -1) const;
+  double get_latest_time() const;
+  double get_oldest_time() const;
+  const PStatFrameData &get_frame_at_time(double time) const;
+  int get_frame_number_at_time(double time, int hint = -1) const;
 
   const PStatFrameData &get_latest_frame() const;
 
   bool get_elapsed_frames(int &then_i, int &now_i) const;
-  float get_frame_rate() const;
+  double get_frame_rate() const;
 
 
-  void set_history(float time);
-  float get_history() const;
+  void set_history(double time);
+  double get_history() const;
 
   void record_new_frame(int frame_number, PStatFrameData *frame_data);
 
@@ -72,7 +72,7 @@ private:
   typedef pdeque<PStatFrameData *> Frames;
   Frames _frames;
   int _first_frame_number;
-  float _history;
+  double _history;
 
   bool _computed_elapsed_frames;
   bool _got_elapsed_frames;
